+++ webroot/manager.html
@@ -54,6 +54,7 @@
 		<div id="pageControls">
 			<button id="prevPageButton" class="pageButton">Previous Page</button>
 			<button id="nextPageButton" class="pageButton">Next Page</button>
+			<button id="clearRecentButton" class="pageButton">Clear Recent</button>
 		</div>
 		<div id="accountControls">
